+++ packages/web3-eth/test/unit/rpc_method_wrappers/get_block_uncle_count.test.ts
@@ -68,7 +68,7 @@ describe('getBlockUncleCount', () => {
 		`should format mockRpcResponse using provided return format\nTitle: %s\nInput parameters: %s\n`,
 		async (_, inputParameters) => {
 			const [inputBlock] = inputParameters;
-			const expectedReturnFormat = { number: FMT_NUMBER.STR, bytes: FMT_BYTES.BUFFER };
+			const expectedReturnFormat = { number: FMT_NUMBER.STR, bytes: FMT_BYTES.UINT8ARRAY };
 			const expectedFormattedResult = format(
 				{ eth: 'uint' },
 				mockRpcResponse,