+++ packages/web3/test/e2e/sepolia/get_transaction_receipt.test.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Bytes } from 'web3-types';
-import { bytesToBuffer, hexToBytes } from 'web3-utils';
+import { bytesToUint8Array, hexToBytes } from 'web3-utils';
 
 import Web3 from '../../../src';
 import { getSystemE2ETestProvider } from '../e2e_utils';
@@ -44,7 +44,7 @@ describe(`${getSystemTestBackend()} tests - getTransactionReceipt`, () => {
 		}>({
 			transactionHash: [
 				'0xe3d28c23ffcd8b94b8f2fd802b4c8dd17a5f19992660acd082ac229ab410b959',
-				bytesToBuffer(
+				bytesToUint8Array(
 					hexToBytes(
 						'0xe3d28c23ffcd8b94b8f2fd802b4c8dd17a5f19992660acd082ac229ab410b959',
 					),