+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -87,4 +87,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed ESM import bugs reported in (#6032) and (#6034)
+-   Fixed ESM import bugs reported in (#6032) and (#6034)
+
+### Changed
+
+-   Replaced `Buffer` for `Uint8Array` (#6004)
+-   The methods `recover`, `encrypt`, `privateKeyToAddress` does not support type `Buffer` but supports type `Uint8Array` (#6004)
+-   The method `parseAndValidatePrivateKey` returns a type `Uint8Array` instead of type `Buffer` (#6004)